oracle_cli/config.py
"""Load VM configuration from config.yaml."""

from pathlib import Path

import yaml

CONFIG_PATH = Path(__file__).resolve().parent.parent / "config.yaml"


def load_config() -> dict:
    with open(CONFIG_PATH) as f:
        return yaml.safe_load(f)


def get_vm_config() -> dict:
    return load_config()["vm"]
