pyproject.toml
[project]
name = "oracle-cli"
version = "0.1.0"
description = "CLI tools for managing Oracle Cloud VM instances"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "fabric>=3.2",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
oci-vm = "oracle_cli.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
